% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs-utils.R
\name{corpus_to_graphs}
\alias{corpus_to_graphs}
\title{\code{corpus_to_graphs} returns a  list of co-ocurrence book graphs from a book Corpus. It's not optimized,
so it might take a while for a large Corpus.}
\usage{
corpus_to_graphs(myCorpus)
}
\arguments{
\item{myCorpus}{a tm VCorpus}
}
\value{
A list of igraph graphs
}
\description{
\code{corpus_to_graphs} returns a  list of co-ocurrence book graphs from a book Corpus. It's not optimized,
so it might take a while for a large Corpus.
}
\examples{
myCorpus <- Corpus(VectorSource("three men wait door say holmes oh indeed
                               seem do thing completely must compliment holmes answer"))
book_graphs <- corpus_to_graphs(myCorpus)
}

