% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs-utils.R
\name{create_edge}
\alias{create_edge}
\title{\code{create_edge} returns an edge for a pair of words in a book graph.}
\usage{
create_edge(pair, node_labels)
}
\arguments{
\item{pair}{A character vector of length 2.}

\item{node_labels}{A character vector of generic length, containing all unique tokens extracted from a text.}
}
\value{
A list of integer vectors of length 2 each, containing the positions of the pair in the token vector.
}
\description{
\code{create_edge} returns an edge for a pair of words in a book graph.
}
\examples{
create_edge(c("three", "men"), c("three", "men", "are", "waiting"))
create_edge(c("three", "are"), c("three", "men", "are", "waiting"))
}

