% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs-utils.R
\name{create_edge_list}
\alias{create_edge_list}
\title{\code{create_edge_list} returns all edges of a book graph.}
\usage{
create_edge_list(corpus_element)
}
\arguments{
\item{book_content}{A character vector containing all words from a text book.}
}
\value{
A list of edge pairs (integer vector of length 2) of the co-ocurrence graph book text.
}
\description{
\code{create_edge_list} returns all edges of a book graph.
}
\examples{
create_edge_list("three men wait door say holmes oh indeed seem do thing completely must compliment holmes answer")
}

