% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs-utils.R
\name{create_graph}
\alias{create_graph}
\title{\code{create_graph} creates a graph from an edge list list}
\usage{
create_graph(edge_list)
}
\arguments{
\item{edge_list}{A list of integer vectors of length 2}
}
\value{
An igraph graph.
}
\description{
\code{create_graph} creates a graph from an edge list list
}
\examples{
create_graph(list(c(1,2), c(2,3), c(3,1)))
}

