% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs-utils.R
\name{create_nodes}
\alias{create_nodes}
\title{\code{create_nodes} returns all nodes for a book graph.}
\usage{
create_nodes(book_content)
}
\arguments{
\item{book_content}{A character vector containing all words from a text book.}
}
\value{
A character vector containing unique tokens from  book.
}
\description{
\code{create_nodes} returns all nodes for a book graph.
}
\examples{
create_nodes("three men wait door say holmes oh indeed seem do thing completely must compliment holmes answer")
}

