% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs-utils.R
\name{get_book_content}
\alias{get_book_content}
\title{\code{get_book_content} returns the content from a book from a VCorpus}
\usage{
get_book_content(corpus_element)
}
\arguments{
\item{corpus_element}{An element from a Vcorpus}
}
\value{
A character vector containing the book content
}
\description{
\code{get_book_content} returns the content from a book from a VCorpus
}
\examples{
myCorpus <- Corpus(VectorSource("three men wait door say holmes oh indeed seem do thing completely must compliment holmes answer"))
#get content from first book in VCorpus
get_book_content(myCorpus[[1]])
}

