% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs-utils.R
\name{get_nodes_labels}
\alias{get_nodes_labels}
\title{\code{get_nodes_labels} returns a  list of nodes labels for each book in a Corpus}
\usage{
get_nodes_labels(myCorpus)
}
\arguments{
\item{myCorpus}{a tm VCorpus}
}
\value{
A list of tokens
}
\description{
\code{get_nodes_labels} returns a  list of nodes labels for each book in a Corpus
}
\examples{
myCorpus <- Corpus(VectorSource("three men wait door say holmes oh indeed
                               seem do thing completely must compliment holmes answer"))
nodes_labels <- get_nodes_labels(myCorpus)
}

