% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmining.R
\docType{package}
\name{gmining}
\alias{gmining}
\alias{gmining-package}
\title{gmining: A package for making data analysis in Project Gutenberg easy.}
\description{
The foo package provides three categories of important functions:
foo, bar and baz.
}
\section{gmining functions}{

The gmining functions ...
}

